Factorize REGISTER_* macros and logic for ports and nifs
esp32, rp2040 and stm32 all use the same logic to use .ctor runtime to
register drivers and nifs collections.

Also use the smaller constructor name for nifs of RP2040 on other platforms to
maintain compatibility as RP2040 builds have to explicitely list nifs so they
are not gc'd. EPS32 is unaffected as enabling of nifs is done through menuconfig.
