+++ src/platforms/esp32/components/avm_sys/CMakeLists.txt
@@ -30,6 +30,7 @@ set(AVM_SYS_COMPONENT_SRCS
     "../../../../libAtomVM/otp_net.c"
     "../../../../libAtomVM/otp_socket.c"
     "../../../../libAtomVM/otp_ssl.c"
+    "../../../../libAtomVM/portnifloader.c"
 )
 
 if (IDF_VERSION_MAJOR GREATER_EQUAL 5)
