+++ src/platforms/rp2040/src/lib/rp2040_sys.h
@@ -35,45 +35,11 @@
 
 #pragma GCC diagnostic pop
 
+#include "portnifloader.h"
 #include "sys.h"
 
 // Because pico sdk uses -gc-section, it is required to add -Wl,-u NAME_nif
 // to make sure a given module nif or port are linked in.
-
-#define REGISTER_PORT_DRIVER(NAME, INIT_CB, DESTROY_CB, CREATE_CB)    \
-    struct PortDriverDef NAME##_port_driver_def = {                   \
-        .port_driver_name = #NAME,                                    \
-        .port_driver_init_cb = INIT_CB,                               \
-        .port_driver_destroy_cb = DESTROY_CB,                         \
-        .port_driver_create_port_cb = CREATE_CB                       \
-    };                                                                \
-                                                                      \
-    struct PortDriverDefListItem NAME##_port_driver_def_list_item = { \
-        .def = &NAME##_port_driver_def                                \
-    };                                                                \
-                                                                      \
-    __attribute__((constructor)) void NAME##register_port_driver()    \
-    {                                                                 \
-        NAME##_port_driver_def_list_item.next = port_driver_list;     \
-        port_driver_list = &NAME##_port_driver_def_list_item;         \
-    }
-
-#define REGISTER_NIF_COLLECTION(NAME, INIT_CB, DESTROY_CB, RESOLVE_NIF_CB)  \
-    struct NifCollectionDef NAME##_nif_collection_def = {                   \
-        .nif_collection_init_cb = INIT_CB,                                  \
-        .nif_collection_destroy_cb = DESTROY_CB,                            \
-        .nif_collection_resolve_nif_cb = RESOLVE_NIF_CB                     \
-    };                                                                      \
-                                                                            \
-    struct NifCollectionDefListItem NAME##_nif_collection_def_list_item = { \
-        .def = &NAME##_nif_collection_def                                   \
-    };                                                                      \
-                                                                            \
-    __attribute__((constructor)) void NAME##_nif()                          \
-    {                                                                       \
-        NAME##_nif_collection_def_list_item.next = nif_collection_list;     \
-        nif_collection_list = &NAME##_nif_collection_def_list_item;         \
-    }
 
 #define EVENT_QUEUE_LEN 16
 
@@ -128,49 +94,5 @@ struct RP2040PlatformData
     mbedtls_ctr_drbg_context random_ctx;
     bool random_is_initialized;
 };
-
-typedef void (*port_driver_init_t)(GlobalContext *global);
-typedef void (*port_driver_destroy_t)(GlobalContext *global);
-typedef Context *(*port_driver_create_port_t)(GlobalContext *global, term opts);
-
-struct PortDriverDef
-{
-    const char *port_driver_name;
-    const port_driver_init_t port_driver_init_cb;
-    const port_driver_destroy_t port_driver_destroy_cb;
-    const port_driver_create_port_t port_driver_create_port_cb;
-};
-
-struct PortDriverDefListItem
-{
-    struct PortDriverDefListItem *next;
-    const struct PortDriverDef *const def;
-};
-
-typedef void (*nif_collection_init_t)(GlobalContext *global);
-typedef void (*nif_collection_destroy_t)(GlobalContext *global);
-typedef const struct Nif *(*nif_collection_resolve_nif_t)(const char *name);
-
-struct NifCollectionDef
-{
-    const nif_collection_init_t nif_collection_init_cb;
-    const nif_collection_destroy_t nif_collection_destroy_cb;
-    const nif_collection_resolve_nif_t nif_collection_resolve_nif_cb;
-};
-
-struct NifCollectionDefListItem
-{
-    struct NifCollectionDefListItem *next;
-    const struct NifCollectionDef *const def;
-};
-
-extern struct PortDriverDefListItem *port_driver_list;
-extern struct NifCollectionDefListItem *nif_collection_list;
-
-void port_driver_init_all(GlobalContext *global);
-void port_driver_destroy_all(GlobalContext *global);
-void nif_collection_init_all(GlobalContext *global);
-void nif_collection_destroy_all(GlobalContext *global);
-const struct Nif *nif_collection_resolve_nif(const char *name);
 
 #endif
